chore(react): move msg(s)ToBotonic to its own file
