+++ packages/botonic-react/src/index.js
@@ -22,4 +22,4 @@ export { ShareButton } from './components/shareButton'
 export { MessageTemplate } from './components/messageTemplate'
 export { customMessage } from './components/customMessage'
 export { BotonicInputTester, BotonicOutputTester } from './botonicTester'
-export { msgToBotonic, msgsToBotonic } from './utils'
+export { msgToBotonic, msgsToBotonic } from './msgToBotonic'
